chore(react): add function-level jsdoc (#221)
this commit adds function-level jsdoc to the react output
wrapper/target. this was done in as a part of an effort to come up to
speed/help others in the same vain.

some type narrowing was performed here, largely to make certain array
types readonly. a conscious decision was made to not make the public
facing api (exposed by src/index.ts) use a ReadonlyArray<T>, so as to
avoid a breaking change (although this may be revisisted in the future)