chore(format): add prettier to project
this commit reworks the pre-existing prettier infrastructure in the
project in an attempt to add continuous integration (ci) checks for
formatting to the project.

prettier and ionic's prettier configuration are added as a top level
dependencies to the project, such that lerna can format all projects
with a 'prettier' script in the future. by using ionic's prettier
configuration, we can converge upon previously agreed upon formatting
config.

ci is updated to run prettier in parallel with other jobs, allowing us
to 'fail-fast' in the event of a formatting error. there is potential
room in the future to refactor ci jobs/steps by adding this parallelism,
but that was not done here in order to limit the scope of this commit.

this commit only handles the angular output target. this allows for
incremental improvement/proof-of-concept to get the idea out to the
teams managing this project for feedback sooner than later