chore(example): commit changes to example project
this commit updates the example project to conform to the latest version
of each output target that is currently installed. the motivation for
this commit is to sync the example project with the current state of the
rest of the repo.

the code changes in this commit were generated by performing the
following operations in the root of the repo:
1. `npm ci`
2. `npm run bootstrap`
3. `npm run build`

this commit does not seek to update any output targets in the example
project, nor does it seek to fix any misconfigurations that may exist in
the example. it only seeks to commit changes that should have been
committed in the past