chore(svelte): remove the output target
this commit removes the svelte output target from the repository. the
ionic framework and stencil teams have collectively agreed that at this
time, we are limited in terms of bandwidth to maintain this output
target. given it's minimal usage, we have put out a call to action (cta)
to the community to see if anyone is interested in picking up this
project and maintaining it themselves.

this output target exists in a small monorepo that also contains the
angular, react, and vue output targets, which ionic wishes to continue
to maintain. as a result, we cannot simply transfer the repository to
someone interested in maintaining the svelte output target. instead, we
will spin the svelte output target out to its own repository
(maintaining git history as best we can). in doing so, we will no longer
need this output target in the repo. as such, we remove it.