feat(deps): require stencil 3.0.0+ for all wrappers (#324)
this commit updates the dependencies of the framework wrappers to
require stencil v3.0.0 or higher (beta, rc releases will no longer be
permitted). stencil v2.9+ is still supported, it is only support for
beta and rc release of stencil v3 that are being dropped.

BREAKING_CHANGE: support for Stencil 3 beta and rc candidates has been
dropped. if using one of those builds, please upgrade to 3.0.0 or higher