chore(repo): add autocomment to triage labeled issues
configure ionitron to automatically label comment any issue that is
labeled 'triage' with the repo's current maintenance status. the
motivation here is to bring clarity/set expectations to the amount of
work that is done in the repo on a day-to-day basis and what its current
goals are.

the workflow that ionitron will follow upon applying this patch is as
follows:
1. a new issue is created
2. ionitron will add the 'triage' label to the issue
3. upon successful application of the triage label, the comment in this
patch will be added to the issue

this pr only adds comments to issues - it does not do the same for
pull requests. at this time, ionitron does not have support for labeling
with a filter on author (e.g. ionic employees, renovatebot). to avoid
unneeded noise in the repo, we choose to only comment on issues for
the time being.